Cleanup, additional test case
